Move to plugin and use @ as relative path
